Readme
BomWeatherApp is using HomeScreen class now. HomeScreen creates all the
stuff in its scene in its' constructor now, display() just does the
anims and shows it. Other classes in the gui package were updated to
make that work. The code that makes the window size animate is in
homescreen at the moment, that might change, but it seems ok there.

Now onto the station lists...