Commenting completed in main branch
There's some serious issues with the GUI, some functions look like they
do the exact same thing as other functions.
Functions aren't used but are left in the code etc.
This would be OK if they were used for unit testing, but it seems to be
not used at all and is just there to waste space.

I believe we are moving over to a new branch to complete our final
product,
however I'm going to cut out some of the gui functions which aren't
necessary because they annoy me.