Culling un-used functions and areas
Moved archived Functions down to the bottom of classes,
Removed unused imports however they could be unused due to the commented
out functions.
Would be easy to check when looking at the functions at a later date.

Everything generally looks nicer now for others to look at and
understand how the program was working in the previous build.