Bugfix ...
Had to make Stage member of PlotWindow instead of parent of PlotWIndow.
This had knock-on effect that restored plot windows wouldn't remove
themselves from list when closed because we were trying to remove the
source of the event from the plotWindows list, but that source no longer
is the plotWindow itself but its' member stage.